feat: implement WidgetProvider & useWidget for accessing specific widget data
