--- widget/embedded/src/Widget.tsx
@@ -1,106 +1,0 @@
-import type { WidgetConfig } from './types';
-import type { WalletType } from '@rango-dev/wallets-shared';
-import type { PropsWithChildren } from 'react';
-
-import { I18nManager, styled } from '@rango-dev/ui';
-import React, { useContext, useEffect, useMemo, useState } from 'react';
-
-import { AppRouter } from './components/AppRouter';
-import { AppRoutes } from './components/AppRoutes';
-import { WidgetEvents } from './components/WidgetEvents';
-import { globalFont } from './globalStyles';
-import { useLanguage } from './hooks/useLanguage';
-import { useTheme } from './hooks/useTheme';
-import QueueManager from './QueueManager';
-import { useAppStore } from './store/AppStore';
-import { useNotificationStore } from './store/notification';
-import { useSettingsStore } from './store/settings';
-import { DEFAULT_CONFIG } from './store/slices/config';
-import { initConfig } from './utils/configs';
-import { WidgetContext, WidgetWallets } from './Wallets';
-
-const MainContainer = styled('div', {
-  width: '100%',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  fontFamily: '$widget',
-  boxSizing: 'border-box',
-  '& *, *::before, *::after': {
-    boxSizing: 'inherit',
-    listStyleType: 'none',
-  },
-  '& *:focus-visible': {
-    outlineColor: '$info500',
-    transition: 'none',
-  },
-});
-
-export type WidgetProps = {
-  config?: WidgetConfig;
-};
-
-export function Main() {
-  globalFont();
-
-  const { fetch: fetchMeta, config } = useAppStore();
-  const { activeTheme } = useTheme(config?.theme || {});
-  const { activeLanguage, changeLanguage } = useLanguage();
-  const [lastConnectedWalletWithNetwork, setLastConnectedWalletWithNetwork] =
-    useState<string>('');
-  const [disconnectedWallet, setDisconnectedWallet] = useState<WalletType>();
-  const widgetContext = useContext(WidgetContext);
-
-  useMemo(() => {
-    if (config?.apiKey) {
-      initConfig({
-        API_KEY: config?.apiKey,
-      });
-    }
-  }, [config]);
-  useEffect(() => {
-    void fetchMeta().catch();
-    void useSettingsStore.persist.rehydrate();
-    void useNotificationStore.persist.rehydrate();
-    widgetContext.onConnectWallet(setLastConnectedWalletWithNetwork);
-  }, []);
-  useEffect(() => {
-    if (config?.language) {
-      changeLanguage(config.language);
-    }
-  }, [config?.language]);
-
-  return (
-    <I18nManager language={activeLanguage}>
-      <MainContainer id="swap-container" className={activeTheme()}>
-        <QueueManager>
-          <WidgetEvents />
-          <AppRouter
-            lastConnectedWallet={lastConnectedWalletWithNetwork}
-            disconnectedWallet={disconnectedWallet}
-            clearDisconnectedWallet={() => {
-              setDisconnectedWallet(undefined);
-            }}>
-            <AppRoutes />
-          </AppRouter>
-        </QueueManager>
-      </MainContainer>
-    </I18nManager>
-  );
-}
-
-export function Widget(props: PropsWithChildren<WidgetProps>) {
-  if (!props.config?.externalWallets) {
-    return (
-      <WidgetWallets
-        config={props.config ?? DEFAULT_CONFIG}
-        providers={props.config?.wallets}
-        options={{
-          walletConnectProjectId: props.config?.walletConnectProjectId,
-        }}>
-        <Main />
-      </WidgetWallets>
-    );
-  }
-  return <Main />;
-}
