+++ widget/embedded/src/containers/App/index.ts
@@ -1,0 +1,1 @@
+export { Main } from './App';
