+++ widget/embedded/src/containers/Wallets/index.ts
@@ -1,0 +1,1 @@
+export { WidgetWallets, WidgetContext } from './Wallets';
