+++ widget/embedded/src/containers/Widget/Widget.types.ts
@@ -1,0 +1,5 @@
+import type { WidgetConfig } from '../../types';
+
+export type WidgetProps = {
+  config?: WidgetConfig;
+};
