+++ widget/embedded/src/containers/Widget/index.ts
@@ -1,0 +1,2 @@
+export { Widget } from './Widget';
+export type { WidgetProps } from './Widget.types';
