+++ widget/embedded/src/containers/WidgetInfo/WidgetInfo.types.ts
@@ -1,0 +1,5 @@
+import type { PendingSwapWithQueueID } from '@rango-dev/queue-manager-rango-preset';
+
+export interface WidgetInfoContextInterface {
+  swaps?: PendingSwapWithQueueID[];
+}
