+++ widget/embedded/src/containers/WidgetInfo/index.ts
@@ -1,0 +1,1 @@
+export { WidgetInfo, useWidget } from './WidgetInfo';
