+++ widget/embedded/src/containers/WidgetProvider/index.ts
@@ -1,0 +1,1 @@
+export { WidgetProvider } from './WidgetProvider';
